pyproject.toml
[project]
name = "smdashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.6.0",
    "dotenv>=0.9.9",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "setuptools>=80.3.1",
    "sqlalchemy>=2.0.40",
    "st-pages>=1.0.1",
    "streamlit>=1.45.0",
    "streamlit-agraph>=0.0.45",
    "watchdog>=6.0.0",
]
